Added auto-removal of temp files

+++ src/kvpaxos/runtests.js
@@ -4,10 +4,17 @@ var execSync = require('child_process').execSync;
 var fs = require('fs');
 var numTest = parseInt(process.argv[2]);
 var removeTemps = process.argv[2];
-execSync('rm *.txt');
+var keepFailed = process.argv[3];
+
+try {
+  execSync('rm *.txt');
+}catch(e){
+}
 
 fs.writeFileSync('master.txt', '');
 var files = [];
+// True if failed. Keep failed only if option available.
+var failedFiles = {};
 var closed = 0;
 for(var i=0;i<numTest;i++){
   var fileName = 'TEST'+ i.toString() + Math.random().toString(26)  + '.txt';
@@ -15,10 +22,17 @@ for(var i=0;i<numTest;i++){
   (function(fileName){
     var test = spawn('go', ['test']);
     test.stdout.on('data', function(data){
+      var stringed = String(data);
+      if(stringed.indexOf('FAIL') >= 0){
+        failedFiles[fileName] = true;
+      }
       fs.appendFile(fileName, data);
     })
     test.stderr.on('data', function(data){
       var stringed = String(data);
+      if(stringed.indexOf('FAIL') >= 0){
+        failedFiles[fileName] = true;
+      }
       if(stringed.indexOf('method') < 0 && stringed.indexOf('die') < 0){
         fs.appendFile(fileName, data);
       }
@@ -34,15 +48,21 @@ for(var i=0;i<numTest;i++){
 
 function combineFiles(files){
   files.forEach(function(el){
-    fs.readFile(el, function(err, data){
-      if(!err){
-        // Append to master logs.
-        fs.appendFile('master.txt', data);
-      }
-      // Remove temp file after done
-      if(removeTemps){
-        exec('rm ' + el);
-      }
-    })
+    if(!keepFailed || (keepFailed && failedFiles[el])){
+      fs.readFile(el, function(err, data){
+        if(!err){
+          // Append to master logs.
+          fs.appendFile('master.txt', data);
+        }
+        // Remove temp file after done
+        if(removeTemps){
+          exec('rm ' + el);
+        }
+      })
+    }
   })
 }
+
+process.on('SIGINT', function(){
+  execSync('rm *.txt');
+})
